Cleanup wasm lib and make SyncState somewhat opaque to js
This introduces a JsSyncState struct to give js as a handle to a real
syncstate which we want to keep in the wasm memory region to avoid lots
of serde stuff.

Some methods (getters and setters) are needed on this struct as the js
tests look inside but these are not technically specified for end users.

Also, now we can keep it mostly opaque we don't need the custom
serializer, yay!