Reduce cloning of jssyncstates in wasm backend
Sync may be called frequently, cloning the sync state a lot thrashes the
garbage collector in js and is a bit pointless in our wasm version.

The only reason for keeping the clone is that the js side may want to
use the old value. This means they can have the previous and the new
one, however it isn't free and I don't think it likely they will be
interested in a previous sync state anyway.

Thus we now take the sync state by value to prevent having to clone it
and then pass it back again in the return list.