Add sendable backend
This introduces both a `Send`able and non-`Send`able backend by sharing a
common `GenericBackend`. This currently just changes the event handler
to require `Send` or not. This is preferred to a mutually exclusive
feature as features should be additive and so both can exist at the same
time (for instance two crates in a workspace with one needing `Send` and
the other not).