Add hash_for_opid function
This function enables users to get the hash that corresponds to the
change that created the operation id given.
This hash can then be used to view historical versions of the document
at that point, useful if multiple things change in a single transaction
and you want to look back in time at those.