rust: Add Automerge::repair
Occasionally implementation bugs mean we produce a saved document which
contains mismatched heads (the heads encoded in the document do not
match the heads which are derived when decompressing the document).
Concrete cases in which we know this has happened are when the
implementation failed to encode the predecessors in the encoded ops, but
_did_ encode the predecessors in the changes and thus the change hashes
didn't match. In this case we should be able to detect the error and
correct it by scanning through the opset looking for ops which have been
`set` by subsequent (as identified by the change graph) changes but
which are not contained in the predecessors of the respective `set` ops.

Implement this in `Automerge::repeair` and expose the functionality in
the CLI tools.