js: Don't copy patchCallback when cloning
The `clone` method which creates a new thread of execution for an
automerge document was copying the patch callback from the old document
and ignoring the `patchCallback` passed in to the `clone` call. Modify
`clone` to use the `patchCallback` passed in and never copy the callback
from the original doc.