Scaffolding for move operation
This is very broken and probably needs a bit of a rethink of how the op
tree works to make it successful:

* Before this change the visiblity of an operation only depends on other
  operations that affect the same location in a map/list.
* After this change the visibility of an operation may differ based on a
  move which may have happened (or not happened) depending on the order
  of other moves in ancestors/descendant nodes.

At a high level I think a better path might be to:
* When inserting a `move` into the tree, also insert a `moved_from` (not
  a delete) that is linked to the `move` in some way (maybe re-using
  `pred`/`succ`, but with the caveat that it's at a different location).
* When resolving moves toggle the visibility of these move/moved_from
  using the rules from the paper to avoid cycles and ensure
  that each 'source' object ends up in exactly one place at a given
  "time" in lamport history.
* Update the visibility checks on ops to check for a visible successor
  (or have toggling the visiblity of the move/move_from ops update succ
  as appropriate).

It also needs some care to avoid making the implementation O(n**2) in
number of moves; likely batching calls to resolve_moves would be
sufficient.

The other thing that needs to change is that currently op tree's have a
`parent`, which will need to become a list of parents at a given
timestamp. You can still get the current parent (and the original
parent) but at a given point in "history" the parent may be in a
different place.

An alternative approach would be to not store the `moved_from` in the
doc explicitly, but generate a placeholder/link from one part of the
op-tree to another when a move is inserted; but it's less clear how that
would work with the op tree as it is today.

Another avenue to consider would be explcitly including the original
location in the document, in addition to the destination and the object
that's moved. I didn't do that because it seemed redundant, and it
doesn't really matter where the object is coming from