Fix regex/tracing-subscriber dependency hell (#582)
Due to a series of unfortunate events involving feature resoution and
the tracing-subscriber and regex crates, the tracing-subscriber
integration in our tests was crashing. This is detailed here:
https://github.com/tokio-rs/tracing/issues/2565

For now we work around this by adding a dev dependency on regex which
enables the required features. tracing-subscriber will be updated
shortly and then we will need to get the updated tracing-subscriber into
test-log and then we can relax.