wasm/js: Run patchCallback in JS, not in webassembly
This fixes #593. Passing the pointer to the current document into the
callback from the webassmebly side is basically never going to be safe
because we end up with multiple borrows of the doc. Even if we were to
not pass the doc pointer into the callback, the user could do something
unsafe by calling another method on a reference to the document retained
elsewhere. Instead, we run the callback in JS, so we don't have to worry
about re-entrant borrows of the document.