Save orphaned changes when saving the whole document (#599)
* Save orphaned changes when saving the whole document

Orphaned changes were being omitted when saving the full document. This
meant that you could end up with a situation where two documents which
have received the same set of messages have different states because one
of them had been saved and reloaded whilst containing an orphaned
change.

Modify Automerge::save() so that it encodes the orphaned changes as
change chunks following the saved document chunk. To allow users to
control this behavior also add `Automerge::save_with_options` and
`AutoCommit::save_with_options` which provide a way to specify that you
wish to discard the orphan changes. Modify `Automerge::load` so that it
no longer throws an error if there are missing dependencies but the
first chunk in the loaded changes was a document chunk.