Remove Expose (#608)
The patch API included a flag on put operations indicating whether this exposed a key which had previously been set but hidden because it was conflicted. Observers were then required to reconstruct the exposed object themselves, which is fiddly and confusing. Remove the expose flag and instead generate the patches required to reconstruct the exposed object so that observers no longer need to think about  this.