New API for mutating lists
The `insertAt` and `deleteAt` methods which are added to the list proxies inside the change function are confusing to both humans and typecheckers because they don't exist on standard arrays. Add top level `insertAt` and `deleteAt` functions which perform the same logic in a more intuitive and remove the `Extend` type which enormously confuses everyone.