Use UTF-8 indices for automerge-c
Currently the C API expects strings to be encoded in UTF-8 (and provides
them that way), but expects programmers to provide indices in UTF-32.

This code changes that so that (by default) indexes are assumed to be
UTF-8, but you can build a version that uses UTF-32 indexing with a
compile time flag.