js: integrate isolated doc in changeAt on empty change (#653)
Context: The `changeAt` function "isolates" the document at some heads
and then executes a change callback on the isolated document. This
allows the user to run a change callback as if it were operating on the
document as at the given heads, rather than whatever the "current" state
is. Once the change callback is complete `changeAt` calls
`automerge_wasm::Automerge::integrate` to move the document back to the
current heads of the document (those that were active before the call to
`changeAt`).

Problem: The logic for calling `integrate` - which is in
`_change` - when the callback was complete did not handle the
case where the callback made no changes at all. This left the document
pointing at the isolated heads even after the `changeAt` call was
complete.

Solution: Modify the `_change` logic to call `integrate` even in the
case of empty changes.