Dont build clocks when making whole-document patches (#655)
Context: `Automerge::diff` allows you to generate patches from any set
of heads to any other set of heads in the document. In the general case
this requires constructing a set of clocks for each set of heads to
determine which ops should be visible in the set of patches. However,
`Automerge::diff` also has some logic to check for cases where we can
skip building clocks in order to generate the patches faster.

Problem: The logic to skip building clocks was incorrectly failing to
skip generating clocks in the case when the `before` heads were empty
and the `after` heads were the heads of the document (i.e. when
generating a patch for the whole current state of the document).

Solution: Set the `PatchLog::heads` clock to `None` in the case of a
full history patch to skip clock generation.