js: Return heads of new changes from changeAt (#672)
`changeAt` is useful because it allows the user to choose when they
synchronize state with the document by saving the heads of the last sync
point and then applying changes since that last sync point using
`changeAt`. This only goes one way however, to go in the other direction
the user needs to be able to obtain the diff from the now applied
changes to the state of the document including any other concurrent
changes. To enable this modify `changeAt` to return a `{newDoc: Doc<T>,
newHeads: Heads | null}` instead of just `Doc<T>`. The returned
`newHeads` is the hash of the transaction created by the change callback
and can be used as the `before` argument to `diff` to obtain a diff from
the result of the changeAt call to the state of the document including
concurrent changes.