javascript: Allow loading a document with missing changes (#698)
The behaviour of `Automerge::load` and all derived APIs has so far been
to throw an error if, having finished loading, we find that there are
changes which have missing dependencies. This is because it seemed like
it would be surprising if you call `load` on a saved file and the
document was empty. There are situations however where you know that
there may be missing dependencies. Add a flag to `load` to indicate that
missing depencies are okay to accomodate this.