@automerge/automerge 2.1.0-alpha.9 (#699)
* Run prettier after yarn install

When releasing new versions of automerge it's not possible to run `yarn
install` in `./javascript` because the latest version of
`automerge-wasm` is not available from NPM yet. This means that the
`js_fmt` CI step fails. Run the format step in the `js_tests` job, which
runs it's own NPM registry (using `javascript/e2e`) first to make sure
that the latest `automerge-wasm` is available.

* @automerge/automerge 2.1.0-alpha.9