Fix SyncState::in_flight logic
## Context

We want `Automerge::generate_sync_message` to only return
`Some` when there is a new message to send. This allows people to call
`generate_sync_message` every time anything has changed and just push
the resulting message down the wire if it exists. This requires us to
track whether we have already generated a message and either a) not
received a message since or b) no local changes have been made. Tracking
(a) is the purpose of `SyncState::in_flight`. This flag is checked
whenever we generate  a sync messaage and cleared
`Automerge::receive_sync_message` whenever the received change both
contains no changes and advertises that the other end has the same heads
as the sender. The idea was that this is a kind of "ack" of the sent
message.

## Problem

The `SyncState::in_flight` flag was not being cleared correctly in
`receive_sync_messaage`. This happened in the following sequence of
events:

* Node 1 generates a sync message and sends it to node 2
* Node 2 receives the sync message, then adds a change locally _which
  triggers a false positive in the sync message from node 1_
* Node 2 generates a sync message and sends it back to Node 1
* Node 1 does not clear the `SyncState::in_flight` flag (more on this
  shortly)

To understand why this happened you need to know a little bit about the
sync protocol. The idea behind the sync protocol is that each message
tells the receiver what changes the sender has and the receiver should
send everything which they have that the sender doesn't have. To do this
each sync message contains three important things:

1. The heads of the document at the last time the sender received a sync
   message from the receiver
2. A bloom filter telling the other end every change that the sender has added
   since the last sync heads
3. A list of change hashes which the receiver explicitly wants

To produce a new message the receiver loads all the changes it has which
are not transitive dependencies of (1) (i.e. the changes since the heads
in the sync message) and then filters out those messages which are in
the bloom filter, then adds any of the explicitly requested changes it
has. This means that we only send changes which the other
end doesn't have. The bloom filter can have false positives, which means
that we might miss some changes the other end needs, but the sender can
explicitly request these changes when it discovers they are missing at
the expenses of an extra round trip.

You can see then that the false positive in the bloom filter which Node
1 initially sent means that Node 2 thinks Node 1 already has the change
it just added and so it doesn't send that change in the response sync
message. This means that when Node 2 receives the sync message it sees
that there are no changes but that the other end has different heads to
it and so does not reset the `in_flight` flag and therefore doesn't
generate the message requesting the missing change. End result, the two
documents appear to be in sync but are in fact not.

## Solution

Just reset `SyncState::in_flight` unconditionally in
`receive_sync_messaage`. The intention behind the complicated logic was
to avoid resending large initial sync messages if the sender generates
new changes before receiving an ack from the other end. This issue
hasn't been observed in practice though and the solution is obviously
incorrect. Let's keep it simple until we know there's a problem to
solve.